pyproject.toml
[tool.poetry]
name = "lesson28"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
psycopg2 = "^2.9.4"
pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
